ENH: Refactor db files and add some docstrings to app.py
